BREAKING: Remove unnecessary API and handle fallback image internally
