+++ packages/houdini-utils/stories/Utils/Fallback.stories.tsx
@@ -3,8 +3,6 @@ import {
   blobify,
   registerPaintWorklet,
   fallbackPaintAnimation,
-  hasMozElement,
-  hasWebkitCanvas,
   PaintWorklet,
   PaintWorkletGeometry,
 } from '@fluentui-contrib/houdini-utils';
@@ -54,9 +52,6 @@ export const fallback = (target: HTMLElement) => {
     duration: '1000ms',
     timingFunction: 'ease-in-out',
     iterationCount: [Infinity],
-    target,
-    onComplete: () => null,
-    onUpdate: () => null,
     delay: '0',
     animations: [
       {
@@ -80,16 +75,6 @@ export const fallback = (target: HTMLElement) => {
   });
 };
 
-const getBackgroundImage = (id: string) => {
-  if (hasMozElement()) {
-    return `-moz-element(#${id})`;
-  } else if (hasWebkitCanvas()) {
-    return `-webkit-canvas(${id})`;
-  }
-
-  return undefined;
-};
-
 const useFallbackAnimation = () => {
   const stateRef = React.useRef<'rest' | 'play'>('rest');
   const playRef = React.useRef<() => void>(() => null);
@@ -101,32 +86,18 @@ const useFallbackAnimation = () => {
       return;
     }
 
-    const { id, play, stop, canvas, cleanup } = fallback(node);
+    const { play, stop, cleanup } = fallback(node);
     stopRef.current = stop;
     cleanupRef.current = cleanup;
-    const backgroundImage = getBackgroundImage(id);
 
     const onComplete = () => {
       stateRef.current = 'rest';
     };
 
-    let onUpdate: () => void = () => null;
-
-    if (backgroundImage) {
-      node.style.backgroundImage = backgroundImage;
-    } else {
-      onUpdate = () => {
-        if (canvas) {
-          const backgroundImage = `url(${canvas.toDataURL('image/png')})`;
-          node.style.backgroundImage = backgroundImage;
-        }
-      };
-    }
-
     playRef.current = () => {
       if (stateRef.current === 'rest') {
         stateRef.current = 'play';
-        play(onComplete, onUpdate);
+        play(onComplete, () => null);
       }
     };
   }, []);
